Remove un-necessary text
Removed ("Note that") because every field in this doc is important.

That copy "Note that" came from eth/caution.md which specify whether the field is same with Ethereum or not.

But in this doc, to summarize important changes I brought fields which have "(Note that)" tags from eth/caution.md